+++ ui/packages/shared/components/src/Dropdown/index.tsx
@@ -65,7 +65,7 @@ const Dropdown = ({
           <Menu.Items
             className={cx(
               width,
-              'absolute w-36 right-0 z-30 mt-2 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none dark:bg-gray-900'
+              'absolute w-36 right-0 z-50 mt-2 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none dark:bg-gray-900'
             )}
           >
             <div className="px-1 py-1 ">{children}</div>