+++ ui/packages/shared/profile/src/ProfileViewWithData.tsx
@@ -43,14 +43,12 @@ interface ProfileViewWithDataProps {
   profileSource: ProfileSource;
   compare?: boolean;
   showVisualizationSelector?: boolean;
-  onSwitchToOneMinute?: () => void;
 }
 
 export const ProfileViewWithData = ({
   queryClient,
   profileSource,
   showVisualizationSelector,
-  onSwitchToOneMinute,
 }: ProfileViewWithDataProps): JSX.Element => {
   const metadata = useGrpcMetadata();
   const [dashboardItems, setDashboardItems] = useURLState<string[]>('dashboard_items', {
@@ -380,7 +378,6 @@ export const ProfileViewWithData = ({
       onDownloadPProf={() => void downloadPProfClick()}
       pprofDownloading={pprofDownloading}
       showVisualizationSelector={showVisualizationSelector}
-      onSwitchToOneMinute={onSwitchToOneMinute}
     />
   );
 };