+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.1021](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1020...@parca/web@0.16.1021) (2026-03-11)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.1020](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1019...@parca/web@0.16.1020) (2026-03-11)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.1019](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1018...@parca/web@0.16.1019) (2026-03-04)
 
 **Note:** Version bump only for package @parca/web