+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.408 (2026-03-11)
+
+**Note:** Version bump only for package @parca/components
+
 ## 0.16.407 (2026-03-03)
 
 # 0.26.0 (2026-02-26)